update style, fix comments
